Copy translation for second string "Only show after sellout of"
This string was added on master while changing the feature of only
showing products after sellout of other products to referring to the
other product instead of a quota. The string for the old functionality
of this feature is still there, but is flagged as "legacy" (the old
feature being deprecated). For the old string a translation had been
added before merging master. This commit copies the translation of the
legacy string to the new string.