Add and improve some translations
Some translations from the GoogleDoc didn't make it into the file,
because the source text accidentally changed in the GoogleDoc (e.g.
"email" -> "e-mail" and one source text got overridden with the
translation). I added those.

I also edited the translation for "Reference code (important)" to make
it consistent with the translation for "Reference code".